+++ mitra_activitypub/src/builders/collection.rs
@@ -22,11 +22,6 @@ pub struct OrderedCollection {
 
     #[serde(skip_serializing_if = "Option::is_none")]
     total_items: Option<i32>,
-
-    // Workaround for Pleroma collection parsing bug
-    // https://git.pleroma.social/pleroma/pleroma/-/issues/3142
-    #[serde(skip_serializing_if = "Option::is_none")]
-    ordered_items: Option<Vec<JsonValue>>,
 }
 
 impl OrderedCollection {
@@ -34,7 +29,6 @@ impl OrderedCollection {
         collection_id: String,
         first_page_id: Option<String>,
         total_items: Option<i32>,
-        with_ordered_items: bool,
     ) -> Self {
         Self {
             _context: build_default_context(),
@@ -42,7 +36,6 @@ impl OrderedCollection {
             object_type: ORDERED_COLLECTION.to_string(),
             first: first_page_id,
             total_items,
-            ordered_items: if with_ordered_items { Some(vec![]) } else { None },
         }
     }
 }