+++ CHANGELOG.md
@@ -17,6 +17,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Process invalid emoji reactions as likes.
 - Renamed identity proof template endpoint to `/api/v1/accounts/identity_claim`.
 
+### Removed
+
+- Removed workaround for Pleroma 2.5 `featured` collection parsing bug.
+
 ## [4.15.0] - 2025-12-22
 
 ### Added